+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Improvements
 
+- Update `elasticsearch-php` dependency to 5.4 [#1648](https://github.com/ruflin/Elastica/pull/1648)
+
 
 ## [5.3.5](https://github.com/ruflin/Elastica/compare/5.3.4...5.3.5)